+++ src/components/common/cardContent.d.ts
@@ -1,0 +1,7 @@
+export interface CardContent {
+  title: string;
+  value?: string | boolean;
+  isRegion?: boolean;
+  linkTo?: string;
+  isAz?: boolean;
+}