+++ src/components/network/NetworkMain.vue
@@ -4,85 +4,20 @@
     <div class="container mt-2">
       <h3>Network resources</h3>
       <div class="row mt-2">
-        <div class="col-12 col-sm-6 col-lg-3 mt-1">
-          <gl-avatar-link to="/network/vpcs">
+        <div
+          class="col-12 col-sm-6 col-lg-3 mt-1"
+          v-for="entry in entries"
+          :key="entry.label"
+        >
+          <gl-avatar-link :to="entry.target">
             <gl-avatar-labeled
-              label="VPCs"
-              sub-label="Isolated Cloud Resources"
-              src="/assets/aws-icons/network/Amazon-VPC.svg"
+              :label="entry.label"
+              :sub-label="entry.subLabel"
+              :src="entry.iconUrl"
               shape="rect"
               :size="32"
             >
             </gl-avatar-labeled>
-          </gl-avatar-link>
-        </div>
-        <div class="col-12 col-sm-6 col-lg-3 mt-1">
-          <gl-avatar-link to="/network/subnets">
-            <gl-avatar-labeled
-              label="Subnets"
-              sub-label="Logical subdivision IP network"
-              src="/assets/aws-icons/network/VPC-subnet-private_light-bg.svg"
-              shape="rect"
-              :size="32"
-            >
-            </gl-avatar-labeled>
-          </gl-avatar-link>
-        </div>
-        <div class="col-12 col-sm-6 col-lg-3 mt-1">
-          <gl-avatar-link to="/network/igws">
-            <gl-avatar-labeled
-              label="Internet Gateways"
-              sub-label="Connect to the Internet"
-              src="/assets/aws-icons/network/Amazon-VPC_Internet-Gateway_light-bg.svg"
-              shape="rect"
-              :size="32"
-            >
-            </gl-avatar-labeled>
-          </gl-avatar-link>
-        </div>
-        <div class="col-12 col-sm-6 col-lg-3 mt-1">
-          <gl-avatar-link to="/network/nats">
-            <gl-avatar-labeled
-              label="Nat Gateways"
-              sub-label="Network Address Translation"
-              src="/assets/aws-icons/network/Amazon-VPC_NAT-Gateway_light-bg.svg"
-              shape="rect"
-              :size="32"
-            >
-            </gl-avatar-labeled>
-          </gl-avatar-link>
-        </div>
-
-        <div class="col-12 col-sm-6 col-lg-3 mt-1">
-          <gl-avatar-link to="/network/routeTables">
-            <gl-avatar-labeled
-              label="Route Tables"
-              sub-label="Route your traffic"
-              src="/assets/aws-icons/network/Amazon-Route-53_Route-Table_light-bg.svg"
-              shape="rect"
-              :size="32"
-            />
-          </gl-avatar-link>
-        </div>
-        <div class="col-12 col-sm-6 col-lg-3 mt-1">
-          <gl-avatar-link to="/network/eips">
-            <gl-avatar-labeled
-              label="Elastic IPs"
-              sub-label="Static IPv4 addresses"
-              src="/assets/aws-icons/network/Amazon-EC2_Elastic-IP-Address_light-bg.svg"
-              shape="rect"
-              :size="32"
-            />
-          </gl-avatar-link>
-        </div>
-        <div class="col-12 col-sm-6 col-lg-3 mt-1">
-          <gl-avatar-link to="/network/securityGroups">
-            <gl-avatar-labeled
-              label="Security groups"
-              sub-label="Virtual firewalls"
-              shape="rect"
-              :size="32"
-            />
           </gl-avatar-link>
         </div>
       </div>
@@ -94,6 +29,7 @@
 import { GlAvatarLabeled, GlAvatarLink } from "@gitlab/ui";
 import { Component, Prop, Vue } from "vue-property-decorator";
 import Header from "@/components/Header/Header.vue";
+import { MenuEntry } from "@/components/common/menuEntry";
 
 @Component({
   components: {
@@ -104,6 +40,59 @@ import Header from "@/components/Header/Header.vue";
 })
 export default class NetworkMain extends Vue {
   @Prop(Boolean) readonly onlyContent: boolean | undefined;
+
+  readonly entries: MenuEntry[] = [
+    {
+      target: "/network/vpcs",
+      label: "VPCs",
+      subLabel: "Isolated Cloud Resources",
+      iconUrl: "/assets/aws-icons/network/Amazon-VPC.svg",
+    },
+    {
+      target: "/network/subnets",
+      label: "Subnets",
+      subLabel: "Logical subdivision IP network",
+      iconUrl: "/assets/aws-icons/network/VPC-subnet-private_light-bg.svg",
+    },
+    {
+      target: "network/igws",
+      label: "Internet Gateways",
+      subLabel: "Connect to the Internet",
+      iconUrl:
+        "/assets/aws-icons/network/Amazon-VPC_Internet-Gateway_light-bg.svg",
+    },
+    {
+      target: "network/nats",
+      label: "Nat Gateways",
+      subLabel: "Network Address Translation",
+      iconUrl: "/assets/aws-icons/network/Amazon-VPC_NAT-Gateway_light-bg.svg",
+    },
+    {
+      target: "network/routeTables",
+      label: "Route Tables",
+      subLabel: "Route your traffic",
+      iconUrl:
+        "/assets/aws-icons/network/Amazon-Route-53_Route-Table_light-bg.svg",
+    },
+    {
+      target: "network/eips",
+      label: "Elastic IPs",
+      subLabel: "Static IPv4 addresses",
+      iconUrl:
+        "/assets/aws-icons/network/Amazon-EC2_Elastic-IP-Address_light-bg.svg",
+    },
+    {
+      target: "network/securityGroups",
+      label: "Security groups",
+      subLabel: "Virtual firewalls",
+    },
+    {
+      target: "network/peeringConnections",
+      label: "Peering connections",
+      subLabel: "Connect VPCs",
+      iconUrl: "/assets/aws-icons/network/Amazon-VPC_Peering_light-bg.svg",
+    },
+  ];
 }
 </script>